requirements.txt
django 
djangorestframework
mysqlclient
djangorestframework-simplejwt